fix: merge develop (adding DB)

+++ src/main/java/com/camping_fisa/bouffonduroiapi/services/questions/CategoryService.java
@@ -23,7 +23,7 @@ public class CategoryService {
     }
 
     public Category updateCategory(long categoryId, UpdateCategoryDTO updateCategoryDTO) {
-        Category existingCategory = categoryRepository.findById(categoryId)
+        Category existingCategory = categoryRepository.findById((int) categoryId)
                 .orElseThrow(() -> new NotFoundException(String.valueOf((int) categoryId), Category.class));
 
         Theme mainTheme = themeRepository.findById(existingCategory.getTheme().getThemeId())
